chore: move OperatorSearchOut schema to v2
